2-lab-apiGatewayIntegration/source/cdk/requirements.txt
aws-cdk.core==1.70.0
aws-cdk.aws-iam==1.70.0
aws-cdk.aws-lambda==1.70.0
aws-cdk.aws-apigateway==1.70.0
aws-cdk.aws-dynamodb==1.70.0